requirements.txt
annotated-types==0.6.0
anyio==4.3.0
backoff==2.2.1
boilerpy3==1.0.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==2.0.2
h11==0.14.0
haystack-ai==2.0.0
haystack-bm25==1.0.2
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lazy-imports==0.3.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.2.0
networkx==3.2.1
numpy==1.26.4
openai==1.14.3
pandas==2.2.1
posthog==3.5.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==2.0.2
